# neo4j_vkg/docs/openapi.yaml
openapi: 3.0.3
info:
  title: Purple Fabric - Enterprise Knowledge Graph Platform API
  description: |
    REST API for the Purple Fabric Enterprise Knowledge Graph Platform.
    Provides document ingestion, ontology management, graph querying, entity resolution,
    multi-tenant workspace management, and AI-powered extraction services.

    ## Authentication
    All endpoints (except `/api/auth/login`, `/api/auth/refresh`, `/api/auth/logout`, and `/api/health`)
    require a Bearer JWT token in the `Authorization` header.

    Access tokens expire in 15 minutes. Use the refresh token endpoint to obtain new tokens.

    ## CSRF Protection
    All state-changing requests (POST, PUT, DELETE) must include the `X-Requested-With` header.

    ## Role-Based Access Control (RBAC)
    Four roles with hierarchical permissions:
    - **viewer** — Read-only access (dashboards, queries, chat)
    - **member** — Upload docs, run extractions, contribute content
    - **manager** — Manage ontologies, clear data, organize workspace
    - **admin** — Full access including user management and system settings

    ## Multi-Tenant Architecture
    Resources are scoped to tenants and workspaces. Pass `tenant_id` and `workspace_id`
    via query parameters or the `X-Tenant-ID` / `X-Workspace-ID` headers.
  version: 1.0.0
  contact:
    name: Purple Fabric Team
servers:
  - url: http://localhost:5002
    description: Local development server
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
...
  schemas:
...
    User:
      type: object
      properties:
        email: { type: string }
        name: { type: string }
        role: { type: string, enum: [viewer, member, manager, admin] }
        workspaces: { type: array, items: { type: string } }
        createdAt: { type: string, format: date-time }
    AuthResponse:
      type: object
      properties:
        token: { type: string, description: JWT access token (15m expiry) }
        refreshToken: { type: string, description: Refresh token (7d expiry) }
        user: { $ref: '#/components/schemas/User' }
...
security:
  - BearerAuth: []
paths:
...
  /api/auth/login:
    post:
      tags: [Auth]
      summary: User login
      security: []
      description: Authenticate with email and password. Returns JWT access token (15m) and refresh token (7d). Rate limited to 15 attempts per 15 minutes per IP. Account locks after 5 failed attempts for 15 minutes.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { description: Invalid credentials }
        '423': { description: Account locked }
        '429': { description: Rate limited }
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [refreshToken]
              properties:
                refreshToken: { type: string }
      responses:
        '200':
          description: New token pair
          content:
            application/json:
              schema: { $ref: '#/components/schemas/AuthResponse' }
        '401': { description: Invalid or expired refresh token }
...
  /api/versioning/documents/{docId}/versions:
    get:
      tags: [Versioning]
      summary: Get extraction versions for a document
      parameters:
        - name: docId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Version list }
  /api/versioning/documents/{docId}/current:
    get:
      tags: [Versioning]
      summary: Get current extraction snapshot
      parameters:
        - name: docId
          in: path
          required: true
          schema: { type: string }
      responses:
        '200': { description: Current snapshot }
  /api/versioning/documents/{docId}/rollback:
    post:
      tags: [Versioning]
      summary: Rollback to a previous extraction version (manager+)
      parameters:
        - name: docId
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [version_id]
              properties:
                version_id: { type: string }
                reason: { type: string }
      responses:
        '200': { description: Rollback completed }
...